Write system tests for authentication features
This is useful for showing that everything works, and for people who are
starting out with Rodauth and want to write basic tests.

Tests weren't working previously due to sequel-activerecord_connection
gem not supporting transaction hooks when Active Record holds the outer
transaction (which is the case in Rails tests). Luckily, this has been
addressed in sequel-activerecord_connection 1.1.0, so now it works
seamlessly just like Active Record's transaction callbacks.