Improve security around downloading recovery codes
Since the download endpoint renders recovery codes without a password
confirmation, the security provided by the required password confirmation
when viewing recovery codes is circumvented, since the attacker can
easily find out the download URL. This allows the attacker to download
recovery codes in case there is an XSS vulnerability and the site
doesn't have strict CSP configured.

We avoid this by modifying the download link to perform the download all
in JavaScript, without a separate endpoint. This ensures recovery codes
can be downloaded only through access to the page, and nowhere else.