Use different authenticated by for remembering MFA
This is more honest, because we haven't actually authenticated via TOTP,
and is consistent with how 1st factor uses the "remember" identifier.
It also doesn't need to be changed when using WebAuthn instead of TOTP.
We're not using "remember" twice to avoid password authentication from
confirm_password feature logging the user out of 2nd factor, as it
deletes all "remember" elements from authenticated by array.