Move Sequel initialization into Rodauth configuration
Some Sequel database extensions require connecting to the database (e.g.
sql_log_normalizer), which might not be always available. For example,
the `db:create` rake task boots the application, but the database is
expected not to exist beforehand. So, we move it to the Rodauth
configuration which is autoloaded.